requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-CORS==4.0.0
Werkzeug==2.3.6
torch==2.6.0
torchvision==0.21.0
transformers==4.30.2
opencv-python==4.8.0.74
Pillow==10.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0

# MongoDB dependencies
mongoengine==3.0.5
pymongo==4.5.0
dnspython==2.4.2
